Changed storage variable to int instead of float

+++ Assets/Scripts/Harvest/HarvestTarget.cs
@@ -5,7 +5,7 @@ namespace RPG.Harvest
 {    
     public class HarvestTarget : MonoBehaviour
     {
-        [SerializeField] float remainingResource = 100;
+        [SerializeField] int remainingResource = 100;
         public bool IsEmpty { get => remainingResource == 0; }
 
         private void Start()
@@ -14,9 +14,9 @@ namespace RPG.Harvest
         }
 
 
-        public float GrabResource(float triedAmount)
+        public int GrabResource(int triedAmount)
         {
-            float grabedAmount = 0;
+            int grabedAmount = 0;
             if (triedAmount > remainingResource)
             {
                 grabedAmount = remainingResource;
